fix
* fix: add event for dump info
* fix: add event for thread warn
* fix: process of thread stop
* fix: file read and write
* refactor: process of runtime
* refactor: structure of priority queue
* test: add concurrency case
* update: fpnn@1.1.7
* style: coce format
* style: readme

+++ Assets/Plugins/rum/platform/RUMFile.cs
@@ -1,7 +1,6 @@
 using System;
 using System.IO;
 using System.Text;
-
 using com.fpnn;
 
 using UnityEngine;
@@ -21,70 +20,49 @@ namespace com.rum {
 
         private static object File_Locker = new object();
 
-        private int _nextIndex;
-        private object index_locker = new object();
-
         private bool _debug;
         private string _secureDataPath;
 
         public RUMFile(int pid, bool debug) {
-
             this._debug = debug;
             this.InitDirectory(String.Format("{0}/{1}{2}", RUMPlatform.SecureDataPath, "rum_events_", pid));
         }
 
         private void InitDirectory (string secureDataPath) {
-
             if (this._debug) {
-
                 Debug.Log("[RUM] local path: " + secureDataPath);
             }
 
             try {
-
                 if (Directory.Exists(secureDataPath) == false) {
-
                     Directory.CreateDirectory(secureDataPath);
-                } 
+                }
 
                 this._secureDataPath = secureDataPath;
             } catch (Exception ex) {
-
                 ErrorRecorderHolder.recordError(ex);
             }
         }
 
         public RUMFile.Result SaveRumLog(int index, byte[] content) {
-
             string path = String.Format("{0}/{1}{2}", this._secureDataPath, FILE_PRE, index);
             return this.WriteFile(path, content);
         }
 
         public RUMFile.Result LoadRumLog() {
+            string name = this.GetEarlierFile(this._secureDataPath);
 
-            string path = null;
-
-            lock (index_locker) {
-
-                int index = 0;
-
-                for (int i = 0; i < 20; i++) {
-
-                    index = this._nextIndex % RUMConfig.FILE_MAX_COUNT;
-                    path = String.Format("{0}/{1}{2}", this._secureDataPath, FILE_PRE, index);
-                    this._nextIndex = ++index;
-
-                    if (new FileInfo(path).Exists) {
-
-                        break;
-                    }
-                }
+            if (string.IsNullOrEmpty(name)) {
+                return new RUMFile.Result() {
+                    success = false,
+                    content = new Exception("fail to load")
+                };
             }
 
+            string path = String.Format("{0}/{1}", this._secureDataPath, name);
             RUMFile.Result res = this.ReadFile(path, true);
 
             if (!res.success) {
-
                 this.DeleteFile(path);
             }
 
@@ -92,18 +70,15 @@ namespace com.rum {
         }
 
         public RUMFile.Result SaveStorage(byte[] content) {
-
             string path = String.Format("{0}/{1}", this._secureDataPath, STORAGE_FILE);
             return this.WriteFile(path, content);
         }
 
         public RUMFile.Result LoadStorage() {
-
             string path = String.Format("{0}/{1}", this._secureDataPath, STORAGE_FILE);
             RUMFile.Result res = this.ReadFile(path, false);
 
             if (!res.success) {
-
                 this.DeleteFile(path);
             }
 
@@ -111,31 +86,23 @@ namespace com.rum {
         }
 
         public RUMFile.Result ClearAllFile() {
-
             return this.DeleteDirectory(this._secureDataPath);
         }
 
         private RUMFile.Result WriteFile(string path, string content, Encoding encoding) {
-
             lock (File_Locker) {
-
                 try {
-
                     using (StreamWriter writer = new StreamWriter(path, false, encoding)) {
-
                         writer.WriteLine(content);
                     }
                 } catch (Exception ex) {
-
                     return new RUMFile.Result() {
-
                         success = false,
                         content = ex
                     };
                 }
 
                 return new RUMFile.Result() {
-
                     success = true,
                     content = content
                 };
@@ -143,26 +110,19 @@ namespace com.rum {
         }
 
         private RUMFile.Result WriteFile(string path, byte[] content) {
-
             lock (File_Locker) {
-
                 try {
-
                     using (FileStream fs = new FileStream(path, FileMode.Create, FileAccess.Write)) {
-
                         fs.Write(content, 0, content.Length);
                     }
                 } catch (Exception ex) {
-
                     return new RUMFile.Result() {
-
                         success = false,
                         content = ex
                     };
                 }
 
                 return new RUMFile.Result() {
-
                     success = true,
                     content = content
                 };
@@ -170,44 +130,34 @@ namespace com.rum {
         }
 
         private RUMFile.Result ReadFile(string path, bool delete, Encoding encoding) {
-
             lock (File_Locker) {
-
                 string content;
 
                 try {
-
                     FileInfo info = new FileInfo(path);
 
                     if (!info.Exists) {
-
                         return new RUMFile.Result() {
-
                             success = false,
                             content = new Exception("no file")
                         };
                     }
 
                     using (StreamReader reader = new StreamReader(info.OpenRead(), encoding)) {
-
                         content = reader.ReadToEnd();
                     }
 
                     if (delete) {
-
                         info.Delete();
                     }
                 } catch (Exception ex) {
-
                     return new RUMFile.Result() {
-
                         success = false,
                         content = ex
                     };
-                } 
+                }
 
                 return new RUMFile.Result() {
-
                     success = true,
                     content = content
                 };
@@ -215,45 +165,35 @@ namespace com.rum {
         }
 
         private RUMFile.Result ReadFile(string path, bool delete) {
-
             lock (File_Locker) {
-
                 byte[] content;
 
                 try {
-
                     FileInfo info = new FileInfo(path);
 
                     if (!info.Exists) {
-
                         return new RUMFile.Result() {
-
                             success = false,
                             content = new Exception("no file")
                         };
                     }
 
                     using (FileStream fs = info.OpenRead()) {
-
                         content = new byte[fs.Length];
                         fs.Read(content, 0, content.Length);
                     }
 
                     if (delete) {
-
                         info.Delete();
                     }
                 } catch (Exception ex) {
-
                     return new RUMFile.Result() {
-
                         success = false,
                         content = ex
                     };
-                } 
+                }
 
                 return new RUMFile.Result() {
-
                     success = true,
                     content = content
                 };
@@ -261,28 +201,21 @@ namespace com.rum {
         }
 
         private RUMFile.Result DeleteFile(string path) {
-
             lock (File_Locker) {
-
                 try {
-
                     FileInfo info = new FileInfo(path);
 
                     if (info.Exists) {
-
                         info.Delete();
                     }
                 } catch (Exception ex) {
-
                     return new RUMFile.Result() {
-
                         success = false,
                         content = ex
                     };
-                } 
+                }
 
                 return new RUMFile.Result() {
-
                     success = true,
                     content = "delete file success"
                 };
@@ -290,29 +223,21 @@ namespace com.rum {
         }
 
         private RUMFile.Result DeleteDirectory(string path) {
-
             lock (File_Locker) {
-
                 try {
-
                     if (Directory.Exists(path)) {
-
                         DirectoryInfo info = new DirectoryInfo(path);
-
                         this.DeleteAllFiles(info);
                         info.Delete();
                     }
-                } catch(Exception ex) {
-
+                } catch (Exception ex) {
                     return new RUMFile.Result() {
-
                         success = false,
                         content = ex
                     };
                 }
 
                 return new RUMFile.Result() {
-
                     success = true,
                     content = "delete directory success"
                 };
@@ -320,20 +245,45 @@ namespace com.rum {
         }
 
         private void DeleteAllFiles(DirectoryInfo info) {
-
             FileInfo[] fis = info.GetFiles();
 
             foreach (FileInfo fi in fis) {
-
                 fi.Delete();
             }
 
             DirectoryInfo[] dis = info.GetDirectories();
 
             foreach (DirectoryInfo di in dis) {
-
                 DeleteAllFiles(di);
                 di.Delete();
+            }
+        }
+
+        private string GetEarlierFile(string path) {
+            lock (File_Locker) {
+                string name = null;
+
+                try {
+                    if (Directory.Exists(path)) {
+                        DirectoryInfo info = new DirectoryInfo(path);
+                        FileInfo[] fis = info.GetFiles();
+
+                        if (fis != null && fis.Length > 1) {
+                            Array.Sort(fis, (x, y) => {
+                                return x.LastWriteTimeUtc.CompareTo(y.LastWriteTimeUtc);
+                            });
+                            name = fis[0].Name;
+
+                            if (STORAGE_FILE == name) {
+                                name = fis[1].Name;
+                            }
+                        }
+                    }
+                } catch (Exception ex) {
+                    ErrorRecorderHolder.recordError(ex);
+                }
+
+                return name;
             }
         }
     }
